requirements.txt
###############################################
# Core Framework
###############################################
fastapi
uvicorn[standard]
python-multipart

###############################################
# Database / ORM / Migrations
###############################################
sqlalchemy>=2.0
psycopg[binary]
alembic
pgvector


###############################################
# DXF / Drawing Processing
###############################################
ezdxf
matplotlib
Pillow


###############################################
# LLM / Embeddings
###############################################
openai>=1.0.0
requests


###############################################
# Settings / Utilities
###############################################
pydantic>=2.7
pydantic-settings>=2.2
python-dotenv


###############################################
# Optional (future, for local LLMs like Ollama)
###############################################
# ollama-python
# or a compatible HTTP client if using a custom wrapper

